search-server/requirements.txt
# Python 3.10+ recommended. Install: pip install -r requirements.txt
# Run: uvicorn main:app --host 127.0.0.1 --port 8000 --reload

python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9
torch>=2.0.0
transformers>=4.40.0
pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
opencv-python-headless>=4.8.0
datasets>=2.16.0
sqlalchemy>=2.0.0
